src/main/kotlin/com/nextbreakpoint/flinkoperator/controller/operation/TaskManagerDetails.kt
package com.nextbreakpoint.flinkoperator.controller.operation

import com.nextbreakpoint.flinkoperator.common.model.ClusterSelector
import com.nextbreakpoint.flinkoperator.common.model.FlinkOptions
import com.nextbreakpoint.flinkoperator.common.model.TaskManagerId
import com.nextbreakpoint.flinkoperator.common.utils.FlinkClient
import com.nextbreakpoint.flinkoperator.common.utils.KubeClient
import com.nextbreakpoint.flinkoperator.controller.core.Operation
import com.nextbreakpoint.flinkoperator.controller.core.OperationResult
import com.nextbreakpoint.flinkoperator.controller.core.OperationStatus
import io.kubernetes.client.JSON
import org.apache.log4j.Logger

class TaskManagerDetails(flinkOptions: FlinkOptions, flinkClient: FlinkClient, kubeClient: KubeClient): Operation<TaskManagerId, String>(flinkOptions, flinkClient, kubeClient) {
    companion object {
        private val logger = Logger.getLogger(TaskManagerDetails::class.simpleName)
    }

    override fun execute(clusterSelector: ClusterSelector, params: TaskManagerId): OperationResult<String> {
        return try {
            val address = kubeClient.findFlinkAddress(flinkOptions, clusterSelector.namespace, clusterSelector.name)

            val details = flinkClient.getTaskManagerDetails(address, params)

            OperationResult(
                OperationStatus.OK,
                JSON().serialize(details)
            )
        } catch (e : Exception) {
            logger.error("[name=${clusterSelector.name}] Can't get details of task manager $params", e)

            OperationResult(
                OperationStatus.ERROR,
                "{}"
            )
        }
    }
}